[builder] Move color layers code to its own file
We are going to add more color layers handling code.
